db-console: move polyfill out of src/js
The tree at pkg/ui/workspaces/db-console/src/js was used almost
exclusively for the @cockroachlabs/crdb-protobuf-client package, with
the exception of one polyfill that's used only during tests. Move the
Object.assign polyfill to .../src/util/polyfills/ so that the
.../src/js/ tree can be completely ignored when not building the
protobuf client.

Release justification: Internal refactor to simplify tests
Release note: None
