+++ pkg/ui/workspaces/db-console/src/util/fetch-mock.ts
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-import "src/js/object-assign";
+import "./polyfills/object-assign";
 import "whatwg-fetch";
 import fetchMock from "fetch-mock";
