+++ pkg/ui/workspaces/db-console/src/util/polyfills/object-assign.js
